Idea for manual coupling context manager
This will have to be worked out properly in the lab, testing what exactly the required commands the Q2 motor controller are.

It will allow us to force the motor back to the original manual coupling coordinate at the end of every run, even if there is an error mid-run.